gluon-radv-filterd: expire routers instead of removing them
For the redirect logic needed for parker, it is beneficial to keep a
list of seen routers and mark them as expired instead of removing them
from the global routers list. This way we don't insert a new redirect
entry for the same router when it shows up again after a mesh link
failure.

All foreach loops are adjusted to correctly ignore expired
routers.